backend/requirements.txt
python
loguru
fastapi
uvicorn-standard
sqlalchemy
psycopg2
alembic